+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/MediaSourceList.java
@@ -30,7 +30,6 @@ import androidx.media3.common.util.HandlerWrapper;
 import androidx.media3.common.util.Log;
 import androidx.media3.common.util.NullableType;
 import androidx.media3.common.util.Util;
-import androidx.media3.datasource.TransferListener;
 import androidx.media3.exoplayer.analytics.AnalyticsCollector;
 import androidx.media3.exoplayer.analytics.PlayerId;
 import androidx.media3.exoplayer.drm.DrmSession;
@@ -46,6 +45,7 @@ import androidx.media3.exoplayer.source.MediaSourceEventListener;
 import androidx.media3.exoplayer.source.ShuffleOrder;
 import androidx.media3.exoplayer.source.ShuffleOrder.DefaultShuffleOrder;
 import androidx.media3.exoplayer.upstream.Allocator;
+import androidx.media3.exoplayer.upstream.BandwidthMeter;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -80,6 +80,7 @@ import java.util.Set;
   private static final String TAG = "MediaSourceList";
 
   private final PlayerId playerId;
+  private final BandwidthMeter bandwidthMeter;
   private final List<MediaSourceHolder> mediaSourceHolders;
   private final IdentityHashMap<MediaPeriod, MediaSourceHolder> mediaSourceByMediaPeriod;
   private final Map<Object, MediaSourceHolder> mediaSourceByUid;
@@ -91,8 +92,6 @@ import java.util.Set;
   private ShuffleOrder shuffleOrder;
   private boolean isPrepared;
 
-  @Nullable private TransferListener mediaTransferListener;
-
   /**
    * Creates the media source list.
    *
@@ -103,13 +102,16 @@ import java.util.Set;
    * @param analyticsCollectorHandler The {@link Handler} to call {@link AnalyticsCollector} methods
    *     on.
    * @param playerId The {@link PlayerId} of the player using this list.
+   * @param bandwidthMeter The {@link BandwidthMeter} of the player using this list.
    */
   public MediaSourceList(
       MediaSourceListInfoRefreshListener listener,
       AnalyticsCollector analyticsCollector,
       HandlerWrapper analyticsCollectorHandler,
-      PlayerId playerId) {
+      PlayerId playerId,
+      BandwidthMeter bandwidthMeter) {
     this.playerId = playerId;
+    this.bandwidthMeter = bandwidthMeter;
     mediaSourceListInfoListener = listener;
     shuffleOrder = new DefaultShuffleOrder(0);
     mediaSourceByMediaPeriod = new IdentityHashMap<>();
@@ -309,9 +311,8 @@ import java.util.Set;
   }
 
   /** Prepares the playlist. */
-  public void prepare(@Nullable TransferListener mediaTransferListener) {
+  public void prepare() {
     checkState(!isPrepared);
-    this.mediaTransferListener = mediaTransferListener;
     for (int i = 0; i < mediaSourceHolders.size(); i++) {
       MediaSourceHolder mediaSourceHolder = mediaSourceHolders.get(i);
       prepareChildSource(mediaSourceHolder);
@@ -473,7 +474,7 @@ import java.util.Set;
     childSources.put(holder, new MediaSourceAndListener(mediaSource, caller, eventListener));
     mediaSource.addEventListener(Util.createHandlerForCurrentOrMainLooper(), eventListener);
     mediaSource.addDrmEventListener(Util.createHandlerForCurrentOrMainLooper(), eventListener);
-    mediaSource.prepareSource(caller, mediaTransferListener, playerId);
+    mediaSource.prepareSource(caller, playerId, bandwidthMeter);
   }
 
   private void maybeReleaseChildSource(MediaSourceHolder mediaSourceHolder) {