+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/source/preload/PreloadMediaSource.java
@@ -16,6 +16,7 @@
 package androidx.media3.exoplayer.source.preload;
 
 import static androidx.media3.common.util.Util.postOrRun;
+import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.base.Preconditions.checkState;
 
@@ -164,6 +165,7 @@ public final class PreloadMediaSource extends WrappingMediaSource {
      * @param preloadLooper The {@link Looper} that will be used for preloading. It should be the
      *     same looper with {@link ExoPlayer.Builder#setPlaybackLooper(Looper)} that will play the
      *     created {@link PreloadMediaSource} instances.
+     * @throws IllegalArgumentException If {@code bandwidthMeter} is {@link BandwidthMeter#NO_OP}.
      */
     public Factory(
         MediaSource.Factory mediaSourceFactory,
@@ -176,6 +178,7 @@ public final class PreloadMediaSource extends WrappingMediaSource {
       this.mediaSourceFactory = mediaSourceFactory;
       this.preloadControl = preloadControl;
       this.trackSelector = trackSelector;
+      checkArgument(bandwidthMeter != BandwidthMeter.NO_OP);
       this.bandwidthMeter = bandwidthMeter;
       this.rendererCapabilities = Arrays.copyOf(rendererCapabilities, rendererCapabilities.length);
       this.loadControl = loadControl;
@@ -311,6 +314,7 @@ public final class PreloadMediaSource extends WrappingMediaSource {
             this.preloadCalled = true;
             this.startPositionUs = startPositionUs;
             setPlayerId(PlayerId.PRELOAD);
+            setBandwidthMeter(bandwidthMeter);
             loadControl.onPrepared(PlayerId.PRELOAD);
             prepareSourceInternal(bandwidthMeter.getTransferListener());
             checkForPreloadError();