+++ libraries/exoplayer/src/test/java/androidx/media3/exoplayer/source/ads/ServerSideAdInsertionMediaSourceTest.java
@@ -74,6 +74,7 @@ import androidx.media3.exoplayer.source.TrackGroupArray;
 import androidx.media3.exoplayer.trackselection.ExoTrackSelection;
 import androidx.media3.exoplayer.trackselection.FixedTrackSelection;
 import androidx.media3.exoplayer.upstream.Allocator;
+import androidx.media3.exoplayer.upstream.BandwidthMeter;
 import androidx.media3.exoplayer.upstream.DefaultAllocator;
 import androidx.media3.test.utils.CapturingAudioSink;
 import androidx.media3.test.utils.DumpFileAsserts;
@@ -153,8 +154,8 @@ public final class ServerSideAdInsertionMediaSourceTest {
 
     mediaSource.prepareSource(
         (source, timeline) -> timelineReference.set(timeline),
-        /* mediaTransferListener= */ null,
-        PlayerId.UNSET);
+        PlayerId.UNSET,
+        BandwidthMeter.NO_OP);
     runMainLooperUntil(() -> timelineReference.get() != null);
 
     Timeline timeline = timelineReference.get();
@@ -231,8 +232,8 @@ public final class ServerSideAdInsertionMediaSourceTest {
         });
     mediaSource.prepareSource(
         (source, timeline) -> timelineReference.set(timeline),
-        /* mediaTransferListener= */ null,
-        PlayerId.UNSET);
+        PlayerId.UNSET,
+        BandwidthMeter.NO_OP);
     runMainLooperUntil(() -> timelineReference.get() != null);
     Timeline firstTimeline = timelineReference.get();
     MediaSource.MediaPeriodId mediaPeriodId1 =
@@ -349,8 +350,8 @@ public final class ServerSideAdInsertionMediaSourceTest {
 
     mediaSource.prepareSource(
         (source, timeline) -> timelineReference.set(timeline),
-        /* mediaTransferListener= */ null,
-        PlayerId.UNSET);
+        PlayerId.UNSET,
+        BandwidthMeter.NO_OP);
     runMainLooperUntil(() -> timelineReference.get() != null);
 
     Timeline timeline = timelineReference.get();
@@ -400,8 +401,8 @@ public final class ServerSideAdInsertionMediaSourceTest {
                 (source, timeline) -> {
                   /* Do nothing. */
                 },
-                /* mediaTransferListener= */ null,
-                PlayerId.UNSET));
+                PlayerId.UNSET,
+                BandwidthMeter.NO_OP));
   }
 
   @Test
@@ -794,9 +795,7 @@ public final class ServerSideAdInsertionMediaSourceTest {
         ImmutableMap.of(periodUid, new AdPlaybackState(/* adsId= */ new Object())), timeline);
     AtomicBoolean sourcePrepared = new AtomicBoolean();
     serverSideAdInsertionMediaSource.prepareSource(
-        (source, newTimeline) -> sourcePrepared.set(true),
-        /* mediaTransferListener= */ null,
-        PlayerId.UNSET);
+        (source, newTimeline) -> sourcePrepared.set(true), PlayerId.UNSET, BandwidthMeter.NO_OP);
     RobolectricUtil.runMainLooperUntil(sourcePrepared::get);
     MediaPeriod serverSideAdInsertionMediaPeriod =
         serverSideAdInsertionMediaSource.createPeriod(