+++ libraries/exoplayer/src/test/java/androidx/media3/exoplayer/source/preload/PreloadMediaSourceTest.java
@@ -283,8 +283,7 @@ public final class PreloadMediaSourceTest {
     PreloadMediaSource preloadMediaSource = preloadMediaSourceFactory.createMediaSource(mediaItem);
 
     MediaSource.MediaSourceCaller externalCaller = mock(MediaSource.MediaSourceCaller.class);
-    preloadMediaSource.prepareSource(
-        externalCaller, bandwidthMeter.getTransferListener(), PlayerId.UNSET);
+    preloadMediaSource.prepareSource(externalCaller, PlayerId.UNSET, bandwidthMeter);
     shadowOf(Looper.getMainLooper()).idle();
     preloadMediaSource.preload(/* startPositionUs= */ 0L);
     shadowOf(Looper.getMainLooper()).idle();
@@ -684,7 +683,7 @@ public final class PreloadMediaSourceTest {
         IllegalStateException.class,
         () ->
             preloadMediaSource.prepareSource(
-                (source, timeline) -> {}, bandwidthMeter.getTransferListener(), PlayerId.UNSET));
+                (source, timeline) -> {}, PlayerId.UNSET, bandwidthMeter));
 
     preloadThread.quit();
   }
@@ -712,9 +711,7 @@ public final class PreloadMediaSourceTest {
         IllegalStateException.class,
         () ->
             preloadMediaSource.prepareSource(
-                mock(MediaSource.MediaSourceCaller.class),
-                bandwidthMeter.getTransferListener(),
-                PlayerId.UNSET));
+                mock(MediaSource.MediaSourceCaller.class), PlayerId.UNSET, bandwidthMeter));
 
     preloadThread.quit();
   }
@@ -740,8 +737,7 @@ public final class PreloadMediaSourceTest {
     preloadMediaSource.preload(/* startPositionUs= */ 0L);
     shadowOf(Looper.getMainLooper()).idle();
     MediaSource.MediaSourceCaller externalCaller = mock(MediaSource.MediaSourceCaller.class);
-    preloadMediaSource.prepareSource(
-        externalCaller, bandwidthMeter.getTransferListener(), PlayerId.UNSET);
+    preloadMediaSource.prepareSource(externalCaller, PlayerId.UNSET, bandwidthMeter);
     wrappedMediaSource.setAllowPreparation(true);
     shadowOf(Looper.getMainLooper()).idle();
 
@@ -770,8 +766,7 @@ public final class PreloadMediaSourceTest {
     preloadMediaSource.preload(/* startPositionUs= */ 0L);
     shadowOf(Looper.getMainLooper()).idle();
     MediaSource.MediaSourceCaller externalCaller = mock(MediaSource.MediaSourceCaller.class);
-    preloadMediaSource.prepareSource(
-        externalCaller, bandwidthMeter.getTransferListener(), PlayerId.UNSET);
+    preloadMediaSource.prepareSource(externalCaller, PlayerId.UNSET, bandwidthMeter);
 
     verify(externalCaller).onSourceInfoRefreshed(eq(preloadMediaSource), any());
     verify(mockPreloadControl).onSourcePrepared(eq(preloadMediaSource));
@@ -797,7 +792,7 @@ public final class PreloadMediaSourceTest {
                         return null;
                       })
                   .when(mockMediaSource)
-                  .prepareSource(any(), any(), any());
+                  .prepareSource(any(), any(), any(BandwidthMeter.class));
               doAnswer(
                       createPeriodInvocation -> {
                         MediaPeriod mediaPeriod = mock(MediaPeriod.class);
@@ -840,8 +835,7 @@ public final class PreloadMediaSourceTest {
     AtomicReference<Timeline> externalCallerSourceInfoTimelineReference = new AtomicReference<>();
     MediaSource.MediaSourceCaller externalCaller =
         (source, timeline) -> externalCallerSourceInfoTimelineReference.set(timeline);
-    preloadMediaSource.prepareSource(
-        externalCaller, bandwidthMeter.getTransferListener(), PlayerId.UNSET);
+    preloadMediaSource.prepareSource(externalCaller, PlayerId.UNSET, bandwidthMeter);
     Pair<Object, Long> periodPosition =
         externalCallerSourceInfoTimelineReference
             .get()
@@ -875,7 +869,7 @@ public final class PreloadMediaSourceTest {
                         return null;
                       })
                   .when(mockMediaSource)
-                  .prepareSource(any(), any(), any());
+                  .prepareSource(any(), any(), any(BandwidthMeter.class));
               doAnswer(
                       createPeriodInvocation -> {
                         MediaPeriod mediaPeriod = mock(MediaPeriod.class);
@@ -918,8 +912,7 @@ public final class PreloadMediaSourceTest {
     AtomicReference<Timeline> externalCallerSourceInfoTimelineReference = new AtomicReference<>();
     MediaSource.MediaSourceCaller externalCaller =
         (source, timeline) -> externalCallerSourceInfoTimelineReference.set(timeline);
-    preloadMediaSource.prepareSource(
-        externalCaller, bandwidthMeter.getTransferListener(), PlayerId.UNSET);
+    preloadMediaSource.prepareSource(externalCaller, PlayerId.UNSET, bandwidthMeter);
     // Create a period from different position.
     Pair<Object, Long> periodPosition =
         externalCallerSourceInfoTimelineReference
@@ -1113,7 +1106,7 @@ public final class PreloadMediaSourceTest {
                         return null;
                       })
                   .when(mockMediaSource)
-                  .prepareSource(any(), any(), any());
+                  .prepareSource(any(), any(), any(BandwidthMeter.class));
               when(mockMediaSource.createPeriod(any(), any(), anyLong()))
                   .thenReturn(mock(MediaPeriod.class));
               return mockMediaSource;
@@ -1133,8 +1126,7 @@ public final class PreloadMediaSourceTest {
     AtomicBoolean externalCallerSourceInfoRefreshedCalled = new AtomicBoolean();
     MediaSource.MediaSourceCaller externalCaller =
         (source, timeline) -> externalCallerSourceInfoRefreshedCalled.set(true);
-    preloadMediaSource.prepareSource(
-        externalCaller, bandwidthMeter.getTransferListener(), PlayerId.UNSET);
+    preloadMediaSource.prepareSource(externalCaller, PlayerId.UNSET, bandwidthMeter);
     shadowOf(Looper.getMainLooper()).idle();
     preloadMediaSource.releaseSource(externalCaller);
 
@@ -1163,7 +1155,7 @@ public final class PreloadMediaSourceTest {
                         return null;
                       })
                   .when(mockMediaSource)
-                  .prepareSource(any(), any(), any());
+                  .prepareSource(any(), any(), any(BandwidthMeter.class));
               when(mockMediaSource.createPeriod(any(), any(), anyLong()))
                   .thenReturn(mock(MediaPeriod.class));
               return mockMediaSource;
@@ -1183,8 +1175,7 @@ public final class PreloadMediaSourceTest {
     MediaSource.MediaSourceCaller externalCaller = mock(MediaSource.MediaSourceCaller.class);
     preloadMediaSource.preload(/* startPositionUs= */ 0L);
     shadowOf(Looper.getMainLooper()).idle();
-    preloadMediaSource.prepareSource(
-        externalCaller, bandwidthMeter.getTransferListener(), PlayerId.UNSET);
+    preloadMediaSource.prepareSource(externalCaller, PlayerId.UNSET, bandwidthMeter);
     preloadMediaSource.releaseSource(externalCaller);
 
     verify(loadControl).onPrepared(PlayerId.PRELOAD);
@@ -1214,7 +1205,7 @@ public final class PreloadMediaSourceTest {
                         return null;
                       })
                   .when(mockMediaSource)
-                  .prepareSource(any(), any(), any());
+                  .prepareSource(any(), any(), any(BandwidthMeter.class));
               when(mockMediaSource.createPeriod(any(), any(), anyLong()))
                   .thenReturn(mock(MediaPeriod.class));
               return mockMediaSource;
@@ -1237,10 +1228,8 @@ public final class PreloadMediaSourceTest {
         (source, timeline) -> externalCaller1SourceInfoRefreshedCalled.set(true);
     MediaSource.MediaSourceCaller externalCaller2 =
         (source, timeline) -> externalCaller2SourceInfoRefreshedCalled.set(true);
-    preloadMediaSource.prepareSource(
-        externalCaller1, bandwidthMeter.getTransferListener(), PlayerId.UNSET);
-    preloadMediaSource.prepareSource(
-        externalCaller2, bandwidthMeter.getTransferListener(), PlayerId.UNSET);
+    preloadMediaSource.prepareSource(externalCaller1, PlayerId.UNSET, bandwidthMeter);
+    preloadMediaSource.prepareSource(externalCaller2, PlayerId.UNSET, bandwidthMeter);
     // Only releaseSource by externalCaller1.
     preloadMediaSource.releaseSource(externalCaller1);
 
@@ -1270,7 +1259,7 @@ public final class PreloadMediaSourceTest {
                         return null;
                       })
                   .when(mockMediaSource)
-                  .prepareSource(any(), any(), any());
+                  .prepareSource(any(), any(), any(BandwidthMeter.class));
               when(mockMediaSource.createPeriod(any(), any(), anyLong()))
                   .thenReturn(mock(MediaPeriod.class));
               return mockMediaSource;
@@ -1317,7 +1306,7 @@ public final class PreloadMediaSourceTest {
                         return null;
                       })
                   .when(mockMediaSource)
-                  .prepareSource(any(), any(), any());
+                  .prepareSource(any(), any(), any(BandwidthMeter.class));
               when(mockMediaSource.createPeriod(any(), any(), anyLong()))
                   .thenReturn(mock(MediaPeriod.class));
               return mockMediaSource;
@@ -1337,8 +1326,7 @@ public final class PreloadMediaSourceTest {
     MediaSource.MediaSourceCaller externalCaller = mock(MediaSource.MediaSourceCaller.class);
     preloadMediaSource.preload(/* startPositionUs= */ 0L);
     shadowOf(Looper.getMainLooper()).idle();
-    preloadMediaSource.prepareSource(
-        externalCaller, bandwidthMeter.getTransferListener(), PlayerId.UNSET);
+    preloadMediaSource.prepareSource(externalCaller, PlayerId.UNSET, bandwidthMeter);
     shadowOf(Looper.getMainLooper()).idle();
     preloadMediaSource.releasePreloadMediaSource();
     shadowOf(Looper.getMainLooper()).idle();