+++ libraries/transformer/src/test/java/androidx/media3/transformer/SpeedProviderMediaPeriodTest.java
@@ -54,6 +54,7 @@ import androidx.media3.exoplayer.source.SinglePeriodTimeline;
 import androidx.media3.exoplayer.source.TrackGroupArray;
 import androidx.media3.exoplayer.trackselection.ExoTrackSelection;
 import androidx.media3.exoplayer.trackselection.FixedTrackSelection;
+import androidx.media3.exoplayer.upstream.BandwidthMeter;
 import androidx.media3.exoplayer.upstream.DefaultAllocator;
 import androidx.media3.test.utils.FakeMediaPeriod;
 import androidx.media3.test.utils.FakeMediaSource;
@@ -421,8 +422,8 @@ public final class SpeedProviderMediaPeriodTest {
           adjustedTimeline.set(timeline);
           latch.open();
         },
-        /* mediaTransferListener= */ null,
-        PlayerId.UNSET);
+        PlayerId.UNSET,
+        BandwidthMeter.NO_OP);
     latch.block(1000);
 
     Window adjustedWindow = adjustedTimeline.get().getWindow(0, new Window());
@@ -460,8 +461,8 @@ public final class SpeedProviderMediaPeriodTest {
           adjustedTimeline.set(timeline);
           latch.open();
         },
-        /* mediaTransferListener= */ null,
-        PlayerId.UNSET);
+        PlayerId.UNSET,
+        BandwidthMeter.NO_OP);
     assertThat(latch.block(1000)).isTrue();
 
     assertThat(adjustedTimeline.get().getWindow(0, new Window()).durationUs).isEqualTo(3_000_000);